[CP-2341] icon refactored
