+++ libs/core/__deprecated__/renderer/components/core/icon/icon-badge-type.constant.ts
@@ -1,0 +1,9 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export enum IconBadgeType {
+  Badge,
+  BadgeWithCounter,
+}
